+++ src/contexts/AuthContext.jsx
@@ -1,36 +1,35 @@
-import React, { createContext, useReducer, useEffect } from "react";
+import React, { createContext, useReducer, useEffect } from 'react';
 import { decodeJWT } from '../helpers/auth-helper';
 
-const getUserFromToken = (token) => {
-    if (!token) return null;
-  
-    const { payload } = decodeJWT(token);
-  
-    return {
-      id: payload.sub,
-      name: payload.name,
-      username: payload.username,
-      roles: payload.roles !== undefined ? payload.roles : null,
-      isAdmin: payload.isAdmin,
-    };
-  };
-
-// Initial state
 const initialState = {
   isAuthenticated: false,
   user: null,
 };
 
+const getUserFromToken = (token) => {
+  if (!token) return null;
+
+  const { payload } = decodeJWT(token);
+
+  return {
+    id: payload.sub,
+    name: payload.name,
+    username: payload.username,
+    roles: payload.roles !== undefined ? payload.roles : null,
+    isAdmin: payload.isAdmin,
+  };
+};
+
 // Reducer function
 const reducer = (state, action) => {
   switch (action.type) {
-    case "LOGIN":
+    case 'LOGIN':
       return {
         ...state,
         isAuthenticated: true,
         user: action.payload,
       };
-    case "LOGOUT":
+    case 'LOGOUT':
       return {
         ...state,
         isAuthenticated: false,
@@ -49,31 +48,31 @@ export const AuthProvider = ({ children }) => {
   const [state, dispatch] = useReducer(reducer, initialState);
 
   useEffect(() => {
-    const token = localStorage.getItem("access_token");
+    const token = localStorage.getItem('access_token');
     if (token) {
       const decoded = getUserFromToken(token);
       dispatch({
-        type: "LOGIN",
+        type: 'LOGIN',
         payload: decoded,
       });
     }
   }, []);
 
   // Actions
-  const login = (data) => {
-    localStorage.setItem("access_token", data.accessToken);
-    localStorage.setItem("refresh_token", data.refreshToken);
+  const loginContext = ({ accessToken, refreshToken }) => {
+    localStorage.setItem('access_token', accessToken);
+    localStorage.setItem('refresh_token', refreshToken);
     dispatch({
-      type: "LOGIN",
-      payload: getUserFromToken(data.accessToken),
+      type: 'LOGIN',
+      payload: getUserFromToken(accessToken),
     });
   };
 
-  const logout = () => {
-    localStorage.removeItem("access_token");
-    localStorage.removeItem("refresh_token");
-    dispatch({ type: "LOGOUT" });
-    location.reload()
+  const logoutContext = () => {
+    localStorage.removeItem('access_token');
+    localStorage.removeItem('refresh_token');
+    dispatch({ type: 'LOGOUT' });
+    location.reload();
   };
 
   return (
@@ -81,8 +80,8 @@ export const AuthProvider = ({ children }) => {
       value={{
         user: state.user,
         isAuthenticated: state.isAuthenticated,
-        login,
-        logout,
+        loginContext,
+        logoutContext,
       }}
     >
       {children}