+++ src/services/httpInterceptor.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { refreshAccessToken } from './authService';
 
 const httpClient = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
@@ -8,19 +9,28 @@ const httpClient = axios.create({
   },
 });
 
+// Helper function to perform token refresh
+const refreshAccessTokenHandler = async (accessToken, refreshToken) => {
+  try {
+    const result = await refreshAccessToken(accessToken, refreshToken);
+    return result.data.accessToken;
+  } catch (error) {
+    console.log(`[refreshAccessToken] Failed to refresh token. ${error}`);
+    return null;
+  }
+};
+
 // Axios request interceptor
 httpClient.interceptors.request.use(
   (config) => {
     // Add current location
     config.lastRequestLocation = window.location.pathname;
 
-    // Add authorization token if needed
-    if (!config.withoutToken) {
-      const token = localStorage.getItem('access_token');
-      if (token) {
-        config.headers['Authorization'] = `Bearer ${token}`;
-      }
+    const token = localStorage.getItem('access_token');
+    if (token) {
+      config.headers['Authorization'] = `Bearer ${token}`;
     }
+
     return config;
   },
   (error) => {
@@ -31,51 +41,57 @@ httpClient.interceptors.request.use(
 // Axios response interceptor
 httpClient.interceptors.response.use(
   (response) => {
-    // Any status code that lie within the range of 2xx cause this function to trigger
     return response;
   },
-  (error) => {
-    // Any status codes that falls outside the range of 2xx cause this function to trigger
-    // Check if the error is due to a network problem or the server is unreachable
-    if (!error.response) {
-      // Handle network errors or server down scenarios
-      console.error('Network error or server is unreachable.');
-      // Optionally, you can redirect to a custom error page or display a global error message
-      window.location = '/error';
-    } else {
-      // Log the HTTP request error details
-      console.error(
-        `HTTP request error: ${error.response.status} ${error.response.statusText} ${error.config.url}`
-      );
-      // Handle specific HTTP status codes
-      if (error.response.status === 401) {
-        // Handle 401 error (unauthorized)
-        let returnUrl = error.config.lastRequestLocation || '/';
+  async (error) => {
+    const originalRequest = error.config;
+    if (
+      error.response &&
+      error.response.status === 401 &&
+      !originalRequest._retry
+    ) {
+      // Try to get a new token
+      const refreshToken = localStorage.getItem('refresh_token');
+      const accessToken = localStorage.getItem('access_token');
+      if (accessToken && refreshToken) {
+        const newAccessToken = await refreshAccessTokenHandler(
+          accessToken,
+          refreshToken
+        );
+        if (newAccessToken) {
+          localStorage.setItem('access_token', newAccessToken);
+          originalRequest._retry = true;
+          originalRequest.headers['Authorization'] = `Bearer ${newAccessToken}`;
+          return httpClient(originalRequest); // Retry the original request with new token
+        } else {
+          // Token refresh failed, redirect to login
+          console.log(
+            `[httpClient.interceptors.response] Error: Token refresh failed, redirect to login`
+          );
+          localStorage.removeItem('access_token');
+          localStorage.removeItem('refresh_token');
+          const returnUrl = originalRequest.lastRequestLocation || '/';
+          window.location = `/login?returnUrl=${encodeURIComponent(returnUrl)}`;
+          return Promise.reject(error);
+        }
+      } else {
+        // No refresh token, redirect to login
+        console.log(
+          `[httpClient.interceptors.response] Error: No refresh token, redirect to login`
+        );
+        const returnUrl = originalRequest.lastRequestLocation || '/';
         window.location = `/login?returnUrl=${encodeURIComponent(returnUrl)}`;
-      } else if (error.response.status === 403) {
-        // Handle 403 error (forbidden)
-        window.location = '/forbidden';
+        return Promise.reject(error);
       }
-      // Additional status code handling can be added here
-
-      /* if (error.response.status === 401 && !originalRequest._retry) {
-      originalRequest._retry = true
-      const refreshToken = localStorageService.getRefreshToken()
-      return axios
-        .post('/auth/token', {
-          refresh_token: refreshToken
-        })
-        .then(res => {
-          if (res.status === 201) {
-            localStorageService.setToken(res.data)
-            axios.defaults.headers.common['Authorization'] =
-              'Bearer ' + localStorageService.getAccessToken()
-            return axios(originalRequest)
-          }
-        })
-    } */
+    } else if (
+      error.response &&
+      error.response.status === 401 &&
+      originalRequest._retry
+    ) {
+      // If the token was already refreshed and still getting 401
+      const returnUrl = originalRequest.lastRequestLocation || '/';
+      window.location = `/login?returnUrl=${encodeURIComponent(returnUrl)}`;
     }
-    // Pass the error to the next error handler
     return Promise.reject(error);
   }
 );