+++ src/services/riskOwnerService.js
@@ -1,0 +1,9 @@
+import httpClient from './httpInterceptor';
+
+const basePath = '/api/risk-owners';
+
+export const getAll = async () => {
+  const { data } = await httpClient.get(`${basePath}`);
+
+  return data;
+};