+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [versionize](https://github.com/versionize/versionize) for commit guidelines.
 
+
+<a name="0.4.0"></a>
+## [0.4.0](https://github.com/opolancoh/proyecta-webapp-react/releases/tag/v0.3.0) (2024-04-23)
+
+### Features
+
+* Support refresh token
+* Add API services
+
+<a name="0.3.0"></a>
+## [0.3.0](https://github.com/opolancoh/proyecta-webapp-react/releases/tag/v0.3.0) (2024-04-16)
+
+### Features
+
+* Add docker containers
+
+
 <a name="0.2.0"></a>
 ## [0.2.0](https://github.com/opolancoh/proyecta-webapp-react/releases/tag/v0.2.0) (2024-04-11)
 
@@ -14,6 +31,7 @@ All notable changes to this project will be documented in this file. See [versio
   - ApiInformation
 * Handle nework issues
 * Improvements and fixes
+
 
 <a name="0.1.0"></a>
 ## [0.1.0](https://github.com/opolancoh/proyecta-webapp-react/releases/tag/v0.1.0) (2024-04-10)